+++ tests/sensors/test_measurements.py
@@ -10,7 +10,6 @@ import pytest
 
 # RESONAATE Imports
 import resonaate.physics.constants as const
-from resonaate.physics.time.stardate import JulianDate
 from resonaate.physics.transforms.eops import EarthOrientationParameter
 from resonaate.physics.transforms.methods import getSlantRangeVector, lla2ecef
 from resonaate.physics.transforms.reductions import updateReductionParameters
@@ -41,7 +40,7 @@ TRUE_EL_RT: float = np.radians(0.01495847759)
 
 # Vallado example 4-1
 LLA: ndarray = np.asarray([np.radians(39.007), np.radians(-104.883), 2.19456])
-JULIAN_DATE: JulianDate = JulianDate.getJulianDate(1994, 5, 14, 13, 11, 20.59856)
+CALENDAR_DATE: datetime.datetime = datetime.datetime(1994, 5, 14, 13, 11, 20, 598560)
 
 # From celestrak.com for May 14, 1994
 # 1994 05 14 49486  0.189443  0.306064 -0.1279402  0.0021743 -0.016163 -0.008660  0.000187  0.000039  28
@@ -61,7 +60,7 @@ EOP: EarthOrientationParameter = EarthOrientationParameter(
 def convertToSEZ(teardown_kvs) -> ndarray:
     """Fixture to get properly converted SEZ observation vector."""
     # pylint: disable=unused-argument
-    updateReductionParameters(JULIAN_DATE, eops=EOP)
+    updateReductionParameters(CALENDAR_DATE, eops=EOP)
     return getSlantRangeVector(lla2ecef(LLA), ECI)